requirements.txt
openai-whisper
spacy
networkx
pyvis

# After installing requirements, also install the Spanish model:
# python -m spacy download es_core_news_sm